refactor: decouple output stats from OutputHandler
