+++ src/agent/implementations/BaseReflectionAgent.ts
@@ -42,6 +42,7 @@ import { ToolState } from '@agent/core/ToolState';
 import type { IModelHandler } from '@agent/modelHandlers';
 import type { ToolDefinition } from '@model';
 import { OutputHandler, NamedOutputFile, IOutputHandler } from '@agent/output';
+import { UsageStatsManager } from '@agent/utils';
 import { messageToSkeleton } from '@agent/utils/messageSkeletonUtils';
 import { BaseAgent } from '@agent/implementations/BaseAgent';
 import { MESSAGE_TYPES } from '@logger/messageTypes';
@@ -67,6 +68,7 @@ export abstract class BaseReflectionAgent extends BaseAgent {
   /** Handler for output file processing and validation. */
   protected outputHandler: IOutputHandler;
   protected latexMediaManager: LatexMediaManager;
+  protected usageStatsManager: UsageStatsManager;
 
   constructor(
     modelHandler: IModelHandler,
@@ -104,13 +106,16 @@ export abstract class BaseReflectionAgent extends BaseAgent {
     this.outputHandler = new OutputHandler(
       this.agentSetting,
       this.agentConfig,
-      this.modelHandler,
       this.logId,
       this.baseFiles,
       this.logger,
     );
 
     this.latexMediaManager = new LatexMediaManager(this.logger);
+    this.usageStatsManager = new UsageStatsManager(
+      this.modelHandler,
+      this.logger,
+    );
   }
 
   /**
@@ -481,7 +486,7 @@ export abstract class BaseReflectionAgent extends BaseAgent {
   /**
    * Processes output files for current round.
    * This method orchestrates the overall output processing flow with clear separation of concerns:
-   * 1. Statistics handling via printStatistics
+   * 1. Statistics handling via UsageStatsManager
    * 2. LaTeX diff operations via handleLatexdiffofOutput (only when endTurn is true)
    *
    * The actual file processing is handled separately in processOutputFiles.
@@ -497,7 +502,11 @@ export abstract class BaseReflectionAgent extends BaseAgent {
     processGroupId?: string,
   ): Promise<string[]> {
     // Print statistics at the end of each round
-    await this.outputHandler.printStatistics(stateGlobal, processGroupId);
+    this.usageStatsManager.reportStatistics(
+      stateGlobal,
+      this.logger.channelId,
+      processGroupId,
+    );
 
     // If this is the end of a turn, handle latexdiff operations as a separate step
     if (
