+++ src/agent/utils/index.ts
@@ -6,3 +6,4 @@ export * from './promptHelpers';
 export * from './outputFileUtils';
 export * from './priceUtils';
 export * from './text';
+export * from './UsageStatsManager';
